+++ enerzyme/models/layers/electron_embedding.py
@@ -5,7 +5,7 @@ from torch import Tensor
 from torch.nn import Linear, init
 import torch.nn.functional as F
 from torch_scatter import segment_sum_coo
-from .mlp import ResidualMLP as _ResidualMLP
+from ..blocks.mlp import ResidualMLP as _ResidualMLP
 from . import BaseFFLayer
 from ..activation import ACTIVATION_KEY_TYPE
